requirements.txt

google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.15.0
google-generativeai==0.8.5
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
streamlit==1.45.1
watchdog==6.0.0
